microscope_select moved out of main
