+++ LS_Pycro_App/main/views/py/__init__.py
@@ -1,5 +1,5 @@
 from PyQt5 import QtWidgets
-from LS_Pycro_App.main.microscope_select.microscope_select import microscope, MicroscopeConfig
+from LS_Pycro_App.microscope_select.microscope_select import microscope, MicroscopeConfig
 
 if microscope == MicroscopeConfig.WILLAMETTE:
     from LS_Pycro_App.main.views.py.wil_main_window import Ui_WilMainWindow as Ui_MainWindow
