+++ LS_Pycro_App/main/views/ui/CONVERT.txt
