+++ LS_Pycro_App/microscope_select/views/py/__init__.py
@@ -1,6 +1,6 @@
 from PyQt5 import QtWidgets
 
-from LS_Pycro_App.main.microscope_select.views.microscope_select_dialog import Ui_MicroscopeSelectDialog
+from LS_Pycro_App.microscope_select.views.py.microscope_select_dialog import Ui_MicroscopeSelectDialog
 
 class MicroscopeSelectDialog(Ui_MicroscopeSelectDialog, QtWidgets.QDialog):
     def __init__(self):
