+++ LS_Pycro_App/microscope_select/views/ui/CONVERT.txt
@@ -1,0 +1,6 @@
+#cmd command to convert ui file to PyQt5 python file.
+
+python -m PyQt5.uic.pyuic "[FILE_PATH].ui" -o "[DESTINATION_PATH].py"  -x
+
+
+
