+++ LS_Pycro_App/microscope_select/views/ui/MicroscopeSelectDialog.ui
