+++ app_start.py
@@ -3,7 +3,7 @@ import sys
 from PyQt5 import QtGui
 from PyQt5.QtWidgets import QApplication
 
-from LS_Pycro_App.main.microscope_select.microscope_select import select_microscope
+from LS_Pycro_App.microscope_select.microscope_select import select_microscope
 
 #__init__ files are run when a package is import for the first time. Thus, we first select the microscope,
 #then the __init__ files select the correct modules to use in their respective namespaces, and then the
